backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
openai
pypdf
python-multipart
tiktoken
